+++ components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swissjs/components
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @swissjs/core@1.2.5
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [740e578]
+  - @swissjs/core@1.2.4
+
 ## 0.1.6
 
 ### Patch Changes